update flowbite theme
